Return LibraryCardInfo instead of BasicLibrary with KeepInfo
