+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/UserProfileCommander.scala
@@ -39,7 +39,8 @@ class UserProfileCommander @Inject() (
     basicUserRepo: BasicUserRepo,
     userConnectionRepo: UserConnectionRepo,
     userValueRepo: UserValueRepo,
-    libraryInfoCommander: LibraryInfoCommander,
+    libraryCardCommander: LibraryCardCommander,
+    libraryMembershipCommander: LibraryMembershipCommander,
     graphServiceClient: GraphServiceClient,
     organizationCommander: OrganizationCommander,
     implicit val defaultContext: ExecutionContext,
@@ -56,8 +57,8 @@ class UserProfileCommander @Inject() (
       val owners = basicUserRepo.loadAll(libOwnerIds)
       val libraryIds = libs.map(_.id.get).toSet
       val memberships = libraryMembershipRepo.getWithLibraryIdsAndUserId(libraryIds, user.id.get)
-      val libraryInfos = libraryInfoCommander.createLibraryCardInfos(libs, owners, user.id, true, idealSize) zip libs
-      val membershipInfos = libs.map { lib => lib.id.get -> memberships(lib.id.get).map(libraryInfoCommander.createMembershipInfo) }.toMap
+      val libraryInfos = libraryCardCommander.createLibraryCardInfos(libs, owners, user.id, true, idealSize) zip libs
+      val membershipInfos = libs.map { lib => lib.id.get -> memberships(lib.id.get).map(libraryMembershipCommander.createMembershipInfo) }.toMap
       (libraryInfos, membershipInfos)
     }
     libraryInfos map {
@@ -101,7 +102,7 @@ class UserProfileCommander @Inject() (
       }
       val libOwnerIds = libs.map(_.ownerId).toSet
       val owners = basicUserRepo.loadAll(libOwnerIds)
-      libraryInfoCommander.createLibraryCardInfos(libs, owners, viewer.flatMap(_.id), true, idealSize)
+      libraryCardCommander.createLibraryCardInfos(libs, owners, viewer.flatMap(_.id), true, idealSize)
     }
   }
 
@@ -123,7 +124,7 @@ class UserProfileCommander @Inject() (
           } else Seq.empty
       }
       val owners = basicUserRepo.loadAll(libs.map(_.ownerId).toSet)
-      libraryInfoCommander.createLibraryCardInfos(libs, owners, viewer.flatMap(_.id), true, idealSize)
+      libraryCardCommander.createLibraryCardInfos(libs, owners, viewer.flatMap(_.id), true, idealSize)
     }
   }
 
@@ -133,7 +134,7 @@ class UserProfileCommander @Inject() (
         val invites = libraryInviteRepo.getByUser(user.id.get, excludeStates = LibraryInviteStates.notActive)
         val libs = invites.sortBy(-_._1.createdAt.getMillis).map(_._2).distinct.drop(page.offset).take(page.limit)
         val owners = basicUserRepo.loadAll((libs.map(_.ownerId)).toSet)
-        libraryInfoCommander.createLibraryCardInfos(libs, owners, viewer.flatMap(_.id), false, idealSize)
+        libraryCardCommander.createLibraryCardInfos(libs, owners, viewer.flatMap(_.id), false, idealSize)
       }
     } else {
       ParSeq.empty
